+++ tests/test_sync.py
@@ -97,6 +97,23 @@ def test_retry_inactive():
     assert 1 == num_called
 
 
+def test_retry_inactive_block():
+    """
+    If inactive, don't retry.
+    """
+    num_called = 0
+
+    stamina.set_active(False)
+
+    with pytest.raises(Exception, match="passed"):  # noqa: PT012
+        for attempt in stamina.retry_context(on=ValueError):
+            with attempt:
+                num_called += 1
+                raise Exception("passed")
+
+    assert 1 == num_called
+
+
 def test_retry_inactive_ok():
     """
     If inactive, the happy path still works.
@@ -111,6 +128,21 @@ def test_retry_inactive_ok():
     stamina.set_active(False)
 
     f()
+
+    assert 1 == num_called
+
+
+def test_retry_inactive_block_ok():
+    """
+    If inactive, the happy path still works.
+    """
+    num_called = 0
+
+    stamina.set_active(False)
+
+    for attempt in stamina.retry_context(on=ValueError):
+        with attempt:
+            num_called += 1
 
     assert 1 == num_called